+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,9 +1,9 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/gavinzzeng/Desktop/Work Space/sdk/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/gavinzzeng/Desktop/Work Space/Oooooori/flutter/cupertino_table_view/example"
+export "FLUTTER_ROOT=/Users/gainvzeng/Desktop/WorkSpace/Flutter/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/gainvzeng/Desktop/WorkSpace/FranklinWH/CupertinoTableView/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
-export "FLUTTER_TARGET=/Users/gavinzzeng/Desktop/Work Space/Oooooori/flutter/cupertino_table_view/example/lib/main.dart"
+export "FLUTTER_TARGET=/Users/gainvzeng/Desktop/WorkSpace/FranklinWH/CupertinoTableView/example/lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
@@ -11,4 +11,4 @@ export "DART_DEFINES=Zmx1dHRlci5pbnNwZWN0b3Iuc3RydWN0dXJlZEVycm9ycz10cnVl,RkxVVF
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"
-export "PACKAGE_CONFIG=/Users/gavinzzeng/Desktop/Work Space/Oooooori/flutter/cupertino_table_view/example/.dart_tool/package_config.json"
+export "PACKAGE_CONFIG=/Users/gainvzeng/Desktop/WorkSpace/FranklinWH/CupertinoTableView/example/.dart_tool/package_config.json"